Move serialization functions from serialization.py to utils.py
The `obj_to_dict` function and related serialization logic have been relocated from `core/serialization.py` to `core/utils.py`. This change consolidates utility functions into a single module and simplifies the codebase by removing unused serialization methods.
